linebot: botと会話ができるようにする

+++ linebot_engine/views.py
@@ -6,11 +6,10 @@ from django.utils.decorators import method_decorator
 from django.views import View
 from django.views.decorators.csrf import csrf_exempt
 from linebot import LineBotApi
-from linebot.models import Message
 
 from linebot_engine.domain.service.line_service import LineService
 from linebot_engine.domain.valueobject.line import WebhookEvent
-from linebot_engine.models import UserProfile
+from linebot_engine.models import UserProfile, Message
 
 WEBHOOK_VERIFICATION_USER_ID = "Udeadbeefdeadbeefdeadbeefdeadbeef"
 
@@ -22,7 +21,7 @@ class CallbackView(View):
         """
         ラインの友達追加時に呼び出され、ラインのIDを登録する
 
-        Notes: LINE DEVELOPERSの画面からWebhookの接続をテストした場合
+        Notes: LINE DEVELOPERS（ビジネスアカウントログイン）の画面からWebhookの接続をテストした場合
           実際のイベント（ユーザーのアクションなど）がないため、eventsデータは存在せず、空の配列が返される
         """
         line_service = LineService()
@@ -52,12 +51,14 @@ class CallbackView(View):
                         )
                     # botがブロックされたとき
                     if event.is_unfollow():
-                        UserProfile.objects.filter(line_user_id).delete()
+                        UserProfile.objects.filter(line_user_id=line_user_id).delete()
 
                     # replyが発生したとき
                     if event.is_message():
                         Message.objects.create(
-                            user_profile=UserProfile.objects.get(user_id=line_user_id),
+                            user_profile=UserProfile.objects.get(
+                                line_user_id=line_user_id
+                            ),
                             source_type=event.event_data.type,
                             message=event.event_data.text,
                         )
